requirements.txt
tensorflow
numpy
inflect==0.2.5
librosa==0.6.0
scipy
tensorboardX
Unidecode==1.0.22
pillow
torch==1.1.0
pysoundfile
